Move daemon tests
